move potentially reusable functions into util
